\name{relMixGUI}
\alias{relMixGUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GUI for relMix
}
\description{
User-friendly graphical user interface for relMix.
}
\usage{
relMixGUI()
}
\details{
Note that if mutation model 'Stepwise' is used, it is required that the names of all non-silent alleles are positive integers. See \code{\link[Familias]{FamiliasLocus}} for details about the mutation models.

Includes error checking for mixture files.
}
\author{
Guro Dorum
}
\seealso{
See \code{\link{relMix}} for the main function implemented in relMixGUI.
}

\examples{
#Example data can be found in the folder "tutorialdata" in the installation folder for relMix
}