% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenos.R
\name{allGenos}
\alias{allGenos}
\title{Find all possible genotypes}
\usage{
allGenos(alleles)
}
\arguments{
\item{alleles}{Vector of input alleles, numeric or character}
}
\value{
Matrix of all possible genotypes, one row per genotype
}
\description{
Finds all possible genotypes based on input alleles.
}
\author{
Guro Dorum
}
