% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relMixGUI.R
\name{relMixGUI}
\alias{relMixGUI}
\title{GUI for relMix}
\usage{
relMixGUI()
}
\description{
User-friendly graphical user interface for relMix.
}
\details{
Includes error checking for the input files.
}
\examples{
#Examples can be found in the vignette and example data files can be found
#in the folder "inst\extdata" in the installation folder for relMix
}
\seealso{
\code{\link{relMix}} for the main function implemented in \code{relMixGUI}.
}
\author{
Guro Dorum, Elias Hernandis
}
